+++ raytrace_defs.cpp
@@ -17,7 +17,7 @@ extern struct raytrace_info rt_info;
 //{{{
 void trace_ray(Color &pixel, const Ray &ray, int depth) {
     Primitive * primitive = NULL;
-    float dist = INF;
+    double dist = INF;
     Scene * scene = Scene::get_instance();
 
     if (depth <= MAX_DEPTH) {
@@ -33,7 +33,7 @@ void trace_ray(Color &pixel, const Ray &ray, int depth) {
             getyx(stdscr, y, x);
             long t = (long)difftime(time(NULL), rt_info.start_time);
             mvprintw(y - 1, 0, "Elapsed time:  %02i:%02i:%02i", t / 3600, (t / 60) % 60, t % 60);
-            mvprintw(y, x + 2, "%02.2f%% done", 100 * (float)rt_info.primary_rays / (float)rt_info.total_primary_rays);
+            mvprintw(y, x + 2, "%02.2f%% done", 100 * (double)rt_info.primary_rays / (double)rt_info.total_primary_rays);
             move(y, x);
             refresh();
         }
@@ -49,8 +49,8 @@ void trace_ray(Color &pixel, const Ray &ray, int depth) {
             pixel += primitive->get_color_contribution(intersection_point, ray, reflect, refract);
             //}}}
             // Reflect the ray, if the surface is reflective.   {{{
-            float reflection_coefficient = primitive->get_reflection();
-            if (reflection_coefficient > 0.0f) {
+            double reflection_coefficient = primitive->get_reflection();
+            if (reflection_coefficient > 0.0) {
                 depth++;
                 trace_ray(pixel, Ray(intersection_point, reflect), depth);
                 depth--;
@@ -70,8 +70,8 @@ void trace_ray(Color &pixel, const Ray &ray, int depth) {
 //{{{
 unsigned long trace_rays(Color * data, Point3D eye) {
     //The image canvas is located around the origin of world space coordinates (0,0,0)
-    float min_x, max_x;
-    float min_y, max_y;
+    double min_x, max_x;
+    double min_y, max_y;
     Scene * scene = Scene::get_instance();
 
     max_x = 15;
@@ -80,16 +80,16 @@ unsigned long trace_rays(Color * data, Point3D eye) {
     min_y = -max_y;
 
     // dx and dy are the amount to add to each pixel to go to the next pixel
-    float dx = (max_x - min_x) / scene->get_viewport_pixel_width();
-    float dy = (max_y - min_y) / scene->get_viewport_pixel_height();
+    double dx = (max_x - min_x) / scene->get_viewport_pixel_width();
+    double dy = (max_y - min_y) / scene->get_viewport_pixel_height();
 
     // Start out in the middle of the top-most left-most pixel.
-    float start_x = min_x + (dx * 0.5);
-    float start_y = min_y + (dy * 0.5);
+    double start_x = min_x + (dx * 0.5);
+    double start_y = min_y + (dy * 0.5);
 
     // This is where the ray will intersect the viewing plane, the window that the 'eye' is
     // looking out of.
-    Point3D screen_intersection(start_x, start_y, 0.0f);
+    Point3D screen_intersection(start_x, start_y, 0.0);
     for (int y = 0; y < scene->get_viewport_pixel_height(); ++y) {
         screen_intersection.x = start_x;
         for (int x = 0; x < scene->get_viewport_pixel_width(); ++x) {
@@ -110,19 +110,19 @@ unsigned long trace_rays(Color * data, Point3D eye) {
 
             // Determine the subpixel dx and dy values, these are used to
             // shift the screen intersection within each pixel
-            float sdx = dx / (2 * scene->get_subpixel_sqrt());
-            float sdy = dy / (2 * scene->get_subpixel_sqrt());
+            double sdx = dx / (2 * scene->get_subpixel_sqrt());
+            double sdy = dy / (2 * scene->get_subpixel_sqrt());
 
             vector<Color> colors;
             // Determine the y-coordinate for the top most subpixel.
-            float sy = screen_intersection.y - (dy * 0.5) + (sdy * 0.5);
+            double sy = screen_intersection.y - (dy * 0.5) + (sdy * 0.5);
             for (int i = 0; i < scene->get_subpixel_sqrt(); i++) {
 
                 // Determine the x-coordinate for the left most subpixel.
-                float sx = screen_intersection.x - (dx * 0.5) + (sdx * 0.5);
+                double sx = screen_intersection.x - (dx * 0.5) + (sdx * 0.5);
                 for (int j = 0; j < scene->get_subpixel_sqrt(); j++) {
-                    float sy_jitter = sy + jitter(sdy);
-                    float sx_jitter = sx + jitter(sdx);
+                    double sy_jitter = sy + jitter(sdy);
+                    double sx_jitter = sx + jitter(sdx);
 
                     Point3D subpixel(sx_jitter, sy_jitter, screen_intersection.z);
                     Ray ray(eye, subpixel - eye);
@@ -137,8 +137,8 @@ unsigned long trace_rays(Color * data, Point3D eye) {
             }
 
             // Average all the colors to get the color value at the window.
-            float red, green, blue;
-            red = green = blue = 0.0f;
+            double red, green, blue;
+            red = green = blue = 0.0;
             vector<Color>::iterator iter, end;
             for (iter = colors.begin(), end = colors.end(); iter != end; iter++) {
                 red += iter->red;
